# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Apployd Control Plane API
  version: 0.1.0
  description: REST API for account, projects, deployments, billing, and observability.
servers:
  - url: http://localhost:4000/api/v1
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create user and organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name, organizationName, organizationSlug]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
                organizationName: { type: string }
                organizationSlug: { type: string }
      responses:
        '201':
          description: Signed up
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Authenticated }
...
  /billing/checkout-session:
    post:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      summary: Start Stripe subscription checkout
      responses:
        '200': { description: Checkout URL }
  /billing/invoices:
    get:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      summary: List recorded invoices
      parameters:
        - in: query
          name: organizationId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Invoice list }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT